requirements.txt
python-bitvavo-api>=1.2.3
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
urllib3>=2.0.0